pyproject.toml
[project]
name = "ask-my-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pinecone>=8.1.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.44.0",
]
